Removing my feeds.
I don't blog frequently enough or have enough content that I want to write
about that is on topic for Planet KDE to continue aggregating my blog or
twitter feeds on planetkde.org

This is unrelated to the unfortunate timing of my last post and is something
I've been considering for awhile. The bad timing just cemented my decision
even further. 

For the curious, I've already sent an apology re: my last post on the planet.


svn path=/trunk/www/sites/planet/; revision=1018650